Continuously update telescope state in observe
- Move telescope state into telescope routes.
- In observe poll the telescope state every 1 s.
