+++ src/telescope_routes.rs
@@ -1,7 +1,9 @@
 use crate::coords::Direction;
-use crate::telescope::{Telescope, TelescopeCollection};
-use crate::telescopes::{ReceiverConfiguration, ReceiverError};
+use crate::telescope::{Telescope, TelescopeCollection, TelescopeContainer};
+use crate::telescopes::{ReceiverConfiguration, ReceiverError, TelescopeStatus};
 use crate::telescopes::{TelescopeError, TelescopeInfo, TelescopeTarget};
+use askama::Template;
+use axum::response::Html;
 use axum::{
     Router,
     extract::{Json, Path, State},
@@ -16,7 +18,8 @@ pub fn routes(telescopes: TelescopeCollection) -> Router {
         .route("/direction", get(get_direction))
         .route("/target", get(get_target).post(set_target))
         .route("/restart", post(restart))
-        .route("/receiver", post(set_receiver_configuration));
+        .route("/receiver", post(set_receiver_configuration))
+        .route("/state", get(get_state));
     Router::new()
         .route("/", get(get_telescopes))
         .nest("/{telescope_id}", telescope_routes)
@@ -39,11 +42,18 @@ async fn get_telescopes(State(telescopes): State<TelescopeCollection>) -> Json<V
 }
 
 #[derive(Debug)]
-struct TelescopeNotFound;
+pub struct TelescopeNotFound;
 
 impl IntoResponse for TelescopeNotFound {
     fn into_response(self) -> Response {
         (StatusCode::NOT_FOUND, "Telescope not found".to_string()).into_response()
+    }
+}
+
+// HACK: Hacky hack!
+impl From<TelescopeError> for TelescopeNotFound {
+    fn from(_: TelescopeError) -> Self {
+        TelescopeNotFound {}
     }
 }
 
@@ -104,4 +114,39 @@ async fn set_receiver_configuration(
 ) -> Result<Json<Result<ReceiverConfiguration, ReceiverError>>, TelescopeNotFound> {
     let mut telescope = extract_telescope(telescopes, telescope_id).await?;
     Ok(Json(telescope.set_receiver_configuration(target).await))
+}
+
+pub async fn get_state(
+    State(telescopes): State<TelescopeCollection>,
+    Path(telescope_id): Path<String>,
+) -> Result<impl IntoResponse, TelescopeNotFound> {
+    let telescopes_lock = telescopes.read().await;
+    let telescope = telescopes_lock
+        .get(&telescope_id)
+        .ok_or(TelescopeNotFound)?;
+    Ok(Html(state(telescope.clone()).await?))
+}
+
+#[derive(Template)]
+#[template(path = "telescope_state.html")]
+struct TelescopeStateTemplate {
+    info: TelescopeInfo,
+    status: String,
+    direction: Direction,
+}
+
+pub async fn state(telescope: TelescopeContainer) -> Result<String, TelescopeNotFound> {
+    let telescope = telescope.telescope.clone().lock_owned().await;
+    let info = telescope.get_info().await?;
+    Ok(TelescopeStateTemplate {
+        info: info.clone(),
+        status: match &info.status {
+            TelescopeStatus::Idle => "Idle".to_string(),
+            TelescopeStatus::Slewing => "Slewing".to_string(),
+            TelescopeStatus::Tracking => "Tracking".to_string(),
+        },
+        direction: telescope.get_direction().await?,
+    }
+    .render()
+    .expect("Template rendering should always succeed"))
 }
